refactor (value): unify use of jdndatepicker directive
