+++ projects/knora-ui/src/lib/viewer/values/jdn-datepicker-directive/jdndatepicker.directive.ts
@@ -2,7 +2,7 @@ import {Directive, Host, Inject, Input, OnChanges} from '@angular/core';
 import {DateAdapter, MAT_DATE_LOCALE} from '@angular/material';
 import {JDNConvertibleCalendar} from 'jdnconvertiblecalendar';
 import {ACTIVE_CALENDAR, JDNConvertibleCalendarDateAdapter} from 'jdnconvertiblecalendardateadapter';
-import {DateInputComponent} from '../date-input/date-input.component';
+import {DateInputComponent} from '../date-value/date-input/date-input.component';
 import {BehaviorSubject} from 'rxjs';
 
 export function makeCalendarToken() {
